Add participant merge conflict handling and harmonization features
- Implemented a new controller for downloading participant merge conflict reports.
- Added functions for managing harmonization decisions and session resolution during participant merges.
- Created UI components for displaying merge summaries, conflicts, and harmonization options.
- Enhanced the participant sourcedata quick select functionality to load files dynamically.
- Updated tests to cover new workflow wiring for participant merge features.
