+++ test_exit.txt
@@ -1,1 +1,1 @@
-0
+1
